% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area-package.R
\docType{package}
\name{area-package}
\alias{area}
\alias{area-package}
\title{area: Calculate Area of Triangles and Polygons}
\description{
Calculate the area of triangles and polygons using the shoestring 
 formula. Area may be signed, taking into account path orientation, or unsigned, 
 ignoring path orientation.
}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com} (\href{https://orcid.org/0000-0002-2471-7511}{ORCID})

}
\keyword{internal}
