% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{area_circle}
\alias{area_circle}
\title{Calculate area given the radius}
\usage{
area_circle(radius = 1)
}
\arguments{
\item{radius}{input parameter is the radius of the circle. The default value is 1}
}
\value{
the area of the circle given radius
}
\description{
To use this package, input the radius of the circle to area_circle(),and output would be the area of the circle
note if the radius is not provided, the default radius=1
}
\examples{
print(area_circle(1))
print(area_circle(2))
}
