% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area-package.R
\docType{data}
\name{mm}
\alias{mm}
\alias{tas}
\alias{mm_tri}
\title{Planar straight line graphs and triangulations}
\description{
A minimal mesh with one hole \code{mm} and a map of Tasmania with multiple
holes in planar straight line graph format from the RTriangle package.
}
\details{
\code{mm_tri} is a triangulated form of \code{mm} in RTriangle \code{triangulation} format.
The \code{H}OLE property is not yet set WIP.
}
\examples{
fs <- list.files(system.file("R", package = "area", mustWork = TRUE))
}
