% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{poly_area}
\alias{poly_area}
\title{Polygon area}
\usage{
poly_area(x, signed = FALSE)
}
\arguments{
\item{x}{polygon in xy matrix}

\item{signed}{defaults to \code{FALSE} and absolute value of area is returned}
}
\description{
Calculate polygon area from a matrix of a closed polygon. Closed means
that the first coordinate is the same as the last.
}
\details{
Currently inputs are not checked whether they are closed polygons or not.

If \code{signed = FALSE} the absolute value of area is returned, otherwise the
sign reflects path orientation. Positive means clockwise orientation.
}
\examples{
x <- c(2, 10, 8, 11, 7, 2)
y <- c(7, 1,  6, 7, 10, 7)
poly_area(cbind(x, y), signed = TRUE)
xy <-
 cbind(x = c(2.3, 1.5, 2.4, 4.5, 4.6, 5.4, 7.6, 8.6, 7.4, 5.1, 2.3),
       y = c(-1.4, 7.3, 22.2, 22.5, 14.4, 11.8, 16.4, 5, 0.8, -1.6, -1.4))
poly_area(xy)
## xy is clockwise so area is negative
poly_area(xy, signed = TRUE)
poly_area(xy[nrow(xy):1, ], signed = TRUE)
}
