% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_by_resolution.R
\name{sort_images}
\alias{sort_images}
\title{Sort images by resolution}
\usage{
sort_images(path, batch = NA)
}
\arguments{
\item{path}{Name of folder in which to sort images.}

\item{batch}{How many images should be put into each sub-folder? Defaults to no sorting.}
}
\description{
Some image-processing programs assume that ones images are sorted by resolution. LS needs to be told what the scale for each image is. It can't guess that. So, I need to read that information in, using exiftool (ie, identify function in Image Magick.) But, it analyses a whole folder at a time. So, I need to look up the resolutions, and move files into folders to be analysed.
}
\examples{
input_dir <- area_example("raw")
sort_images(input_dir)
sort_images(input_dir, 2)
}
