% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangles.R
\name{tri_area}
\alias{tri_area}
\title{Polygon area}
\usage{
tri_area(x, signed = FALSE)
}
\arguments{
\item{x}{coordinates x,y in triplets matrix where nrow(x) = ntriangles*3}

\item{signed}{defaults to \code{FALSE} and absolute value of area is returned}
}
\description{
Calculate polygon area from a matrix of a closed polygon. Closed means
that the first coordinate is the same as the last.
}
\details{
Currently inputs are not checked whether they are closed polygons or not.

If \code{signed = FALSE} the absolute value of area is returned, otherwise the
sign reflects path orientation. Positive means counter-clockwise orientation.
}
\examples{
tri_area(mm_tri$P[t(mm_tri$T), ])
}
