% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{area_circle}
\alias{area_circle}
\title{Calculate the area of a circle given its radius}
\usage{
area_circle(radius = 1)
}
\arguments{
\item{radius}{Input parameter is radius of circle. Default radius is 1.}
}
\value{
Area of circle with radius given
}
\description{
To use this function, call area_circle(r), where r is radius of your desired circle.
Note that if r is not provided, default value of r = 1.
}
\examples{
Area1 = area_circle()
Area2 = area_circle(2)

}
